job sem, move results to new file
