+++ toolkit/tools/pkg/licensecheck/licensecheck.go
@@ -37,7 +37,6 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/microsoft/azurelinux/toolkit/tools/internal/jsonutils"
 	"github.com/microsoft/azurelinux/toolkit/tools/internal/logger"
 	"github.com/microsoft/azurelinux/toolkit/tools/internal/rpm"
 	"github.com/microsoft/azurelinux/toolkit/tools/internal/sliceutils"
@@ -50,64 +49,13 @@ const (
 	numParallelJobs = 20
 )
 
-var jobSemaphore = make(chan struct{}, numParallelJobs) // Limit the number of parallel jobs
-
-// LicenseCheckResult is the result of a license check on an single RPM
-type LicenseCheckResult struct {
-	RpmPath        string   `json:"RpmPath"`
-	PackageName    string   `json:"PackageName,omitempty"`
-	BadDocs        []string `json:"BadDocs,omitempty"`
-	BadFiles       []string `json:"BadFiles,omitempty"`
-	DuplicatedDocs []string `json:"DuplicatedDocs,omitempty"`
-	err            error    `json:"-"`
-}
-
 // LicenseChecker is a tool for searching RPMs for bad licenses
 type LicenseChecker struct {
 	simpleToolChroot *simpletoolchroot.SimpleToolChroot // The chroot to scan the RPMs in
 	distTag          string                             // The distribution tag to use when parsing RPMs
 	exceptions       LicenseExceptions                  // Files that should be ignored
 	results          []LicenseCheckResult               // The results of the search
-}
-
-// HasBadResult returns true if the result contains at least one bad document or file.
-func (r *LicenseCheckResult) HasBadResult() bool {
-	return len(r.BadDocs) > 0 || len(r.BadFiles) > 0
-}
-
-// HasWarningResult returns true if the result contains at least one duplicated documents.
-func (r *LicenseCheckResult) HasWarningResult() bool {
-	return len(r.DuplicatedDocs) > 0
-}
-
-// SaveLicenseCheckResults saves a list of all warnings and errors to a json file.
-func SaveLicenseCheckResults(savePath string, resultsList []LicenseCheckResult) error {
-	// Create parent dir if missing
-	err := os.MkdirAll(filepath.Dir(savePath), os.ModePerm)
-	if err != nil {
-		return fmt.Errorf("failed to create directory for results file. Error:\n%w", err)
-	}
-
-	sortedList := sortAndFilterResults(resultsList)
-	err = jsonutils.WriteJSONFile(savePath, sortedList)
-	if err != nil {
-		return fmt.Errorf("failed to save license check results. Error:\n%w", err)
-	}
-	return nil
-}
-
-// sortAndFilterResults returns a copy of the list that is sorted and contains only warnings and errors.
-func sortAndFilterResults(results []LicenseCheckResult) (sortedList []LicenseCheckResult) {
-	sortedList = make([]LicenseCheckResult, len(results))
-	for i, result := range results {
-		if result.HasBadResult() || result.HasWarningResult() {
-			sortedList[i] = result
-		}
-	}
-	sort.Slice(sortedList, func(i, j int) bool {
-		return sortedList[i].RpmPath < sortedList[j].RpmPath
-	})
-	return sortedList
+	jobSemaphore     chan struct{}                      // Limit the number of parallel jobs
 }
 
 // New creates a new license checker. If this returns successfully the caller is responsible for calling CleanUp().
@@ -120,6 +68,7 @@ func New(buildDirPath, workerTarPath, rpmDirPath, exceptionFilePath, distTag str
 	newLicenseChecker = &LicenseChecker{
 		distTag:          distTag,
 		simpleToolChroot: &simpletoolchroot.SimpleToolChroot{},
+		jobSemaphore:     make(chan struct{}, numParallelJobs),
 	}
 
 	err = newLicenseChecker.simpleToolChroot.InitializeChroot(buildDirPath, chrootName, workerTarPath, rpmDirPath)
@@ -333,12 +282,12 @@ func (l *LicenseChecker) queueWorkers(rpmsToSearchPaths []string, resultsChannel
 		go func(rpmPath string) {
 			// Wait for the semaphore, or allow cancel before running
 			select {
-			case jobSemaphore <- struct{}{}:
+			case l.jobSemaphore <- struct{}{}:
 			case <-cancel:
 				return
 			}
 			defer func() {
-				<-jobSemaphore
+				<-l.jobSemaphore
 			}()
 
 			logger.Log.Debugf("Searching (%s)", filepath.Base(rpmPath))
